Add CLI output sorting
Requires a bit of a refactor, which stinks a tad. Was wrong in the first
place to nest multiple results for the same rule under the same object
anyway.

We have to do this here rather than in the calling shell, à la sort -k,
because most can't parse logfmt and break on fields rather than naive
spaces, and the print format doesn't emit the names of the columns to
sort by anyway, which aren't always at the same position.