Replace all uses of table_schema with schema_name
Because we'll have to lint views, procedures, schemas themselves, etc.,
and there's no point in having different names. In fact, it's a bit of
an inconsistency in information_schema unless you have two schema-able
objects in the same infoschema view, such as a target unique
constraint's schema in an fkey and the FK itself's schema name.